[wip] working queue mode and regular mode - we attempt to pull tests from API before we call build_distributions#last
