+++ lib/knapsack_pro.rb
@@ -58,6 +58,7 @@ require_relative 'knapsack_pro/adapters/spinach_adapter'
 require_relative 'knapsack_pro/allocator'
 require_relative 'knapsack_pro/queue_allocator'
 require_relative 'knapsack_pro/mask_string'
+require_relative 'knapsack_pro/test_suite_builder'
 require_relative 'knapsack_pro/test_case_mergers/base_merger'
 require_relative 'knapsack_pro/test_case_mergers/rspec_merger'
 require_relative 'knapsack_pro/build_distribution_fetcher'
