+++ lib/knapsack_pro/crypto/decryptor.rb
@@ -5,9 +5,9 @@ module KnapsackPro
     class Decryptor
       class TooManyEncryptedTestFilesError < StandardError; end
 
-      def self.call(test_files, encrypted_test_files)
+      def self.call(test_suite_builder, encrypted_test_files)
         if KnapsackPro::Config::Env.test_files_encrypted?
-          new(test_files, encrypted_test_files).call
+          new(test_suite_builder.fast_and_slow_test_files_to_run, encrypted_test_files).call
         else
           # those test files are not encrypted
           encrypted_test_files
